+++ rag-toolchain/src/common/mod.rs
@@ -1,2 +1,5 @@
-pub mod embedding_shared;
-pub mod types;
+mod embedding_shared;
+mod types;
+
+pub use embedding_shared::*;
+pub use types::*;
