+++ rag-toolchain/src/loaders/mod.rs
@@ -1,2 +1,6 @@
-pub mod single_file_loader;
-pub mod traits;
+mod single_file_loader;
+mod traits;
+
+pub use single_file_loader::SingleFileSource;
+pub use traits::AsyncLoadSource;
+pub use traits::LoadSource;
