+++ rag-toolchain/src/retrievers/mod.rs
@@ -1,2 +1,5 @@
-pub mod postgres_vector_retriever;
-pub mod traits;
+mod postgres_vector_retriever;
+mod traits;
+
+pub use postgres_vector_retriever::{PostgresRetrieverError, PostgresVectorRetriever};
+pub use traits::AsyncRetriever;
