+++ rag-toolchain/src/stores/traits.rs
@@ -1,4 +1,4 @@
-use crate::common::types::{Chunk, Embedding};
+use crate::common::{Chunk, Embedding};
 use async_trait::async_trait;
 use std::error::Error;
